add fake seed data

+++ bike-traffic-api/db/migrate/20200614063320_create_bike_trails.rb
@@ -2,7 +2,7 @@ class CreateBikeTrails < ActiveRecord::Migration[6.0]
   def change
     create_table :bike_trails do |t|
       t.string :name
-      t.integer :miles
+      t.float :miles
       t.string :trail_type
       t.integer :bike_trails_counters_id
